Update patch set 1
Patch Set 1:

I wasn't sure if I should do it for libartd.so too. I did run the numbers for the libartd.so and here they are:

arm  ~698K
arm64 ~698K
mips ~466K
x86 ~847K
x86_64 ~847K

So if people would like me to enable this for that library too, I can. I was mostly worried that it would too big to have both inflated by these amounts, but that fear is probably unfounded.

Patch-set: 1