Update patch set 1
Patch Set 1:

Consider:
https://android-review.googlesource.com/#/c/83117/2/compiler/dex/ir_structures.h

If you look at ChildBlockIterator, it contains a pointer to the successors iterator. I had no better way to do it except do it by hand instead.

Using an iterator on the heap was the easiest in this case (though this is an initial version, I have changes and less comments now ;-))

Patch-set: 1