Update patch set 1
Patch Set 1: Code-Review+2

Looks like a clear improvement.  But I'm concerned about the reader side of this.  The problem is that the load of the class status and the resulting load of a (supposedly initialized) field may be reordered, causing the reader to see uninitialized data with a status indicating that it has been initialized.

I'm fine with checking this in, but I think we also need the other half.

Patch-set: 1
Label: Code-Review=+2