Update patch set 3
Patch Set 3:

> Andreas, I'm pretty sure. Your changes are significant enough to
 > make it a new test. But it might hide some undiagnosed failures on
 > the original one, which should not be crashing in the first place.

So I wrote this test, I'm not really concerned with this refactoring. We've always been adding more to this, it really is not intended to be a simple unit test, but something to have interactions between different but overlapping areas. I was the one who requested to the new case to be added. I'm not offended that Andreas thinks my old code is painful.

Patch-set: 3