Update patch set 3
Patch Set 3:

> > Andreas, I'm pretty sure. Your changes are significant enough to
 > > make it a new test. But it might hide some undiagnosed failures
 > on
 > > the original one, which should not be crashing in the first
 > place.
 > 
 > So I wrote this test, I'm not really concerned with this
 > refactoring. We've always been adding more to this, it really is
 > not intended to be a simple unit test, but something to have
 > interactions between different but overlapping areas. I was the one
 > who requested to the new case to be added. I'm not offended that
 > Andreas thinks my old code is painful.

I'm not complaining about the rewrite, on the contrary. What I want to make sure is that we don't forget about the current flakiness. If we replace the current one with this one, which ends up not being flaky, I would not be surprised if after a while we considered the bug as fixed because not showing up anymore... Hence the reason for keeping the current one until we can diagnose and fix the bug.

Patch-set: 3
Label: Verified=0