Update patch set 1
Patch Set 1:

A32 and T32 are the ways the ARMv8 docs refer to the ARMv7 32-bit instruction set. A64 is the name of the new 64-bit ARMv8 instruction set. Therefore, I agree with Zheng, A32_ is not a good prefix for A64 register names. This patch looks good to me.

Patch-set: 1