Update patch set 1
Patch Set 1:

The other tests are golden tests, so they need their own pattern (that is the expected file).

The only thing we could remove from them is the src/ directory. But sometimes we want to play with src/ and src2/, so keeping as is works as well.

The mere point of this CL is to make it faster and easier to write tests. So less excuses to avoid one.

Patch-set: 1