Update patch set 1
Patch Set 1:

Aside from non-deterministic output and using a exit code to confer success, the oat tests do an explicit dex2oat rather than relying on the runtime to fork it. This can be useful for debugging dex2oat and so we should probably make it a run-test option. Note that the dex2oat in the oat tests is brittle and if the generated file has the wrong name the run-time will silently compile the file again. In the worst instance this can make you think you are testing an optimizing compiled thing whereas the code was really compiled with quick.

Patch-set: 1