Update patch set 1
Patch Set 1:

> I don't see why this test needs a change in the test
 > infrastructure. Having a 3rd way of writing tests, seems more
 > confusing. Can you clarify why just adding an empty expected.txt
 > doesn't work?

This test does not need a change. It's an example for where I'd like us to be.

This new way makes a few steps less to write tests. By experience, the less steps to writing the tests, the more tests we'll get. Not doing so leads to things like https://android.googlesource.com/platform/art/+/master/test/083-compiler-regressions/src/Main.java, where it's easier to just add to an existing test rather than create a new file. With such practice, one failure in that multi-test makes it so that we're loosing coverage on the remaining tests in the file.

Patch-set: 1