Update patch set 1
Patch Set 1:

> If there is a test, whether it appears as part of 1 file or as a
 > stand-alone test, its still just the contribution of 1 test. Adding
 > files doesn't bring about extra test coverage.

When I mentioned coverage, I talked about running multiple tests in the same test. So if you have one failure in that test, you won't run the remaining ones.

 My concern in making
 > every bug a test is that we're going to explode the number of test
 > cases to the point that folks won't run as many tests.

Yeah, I share that concern as well.

So there's two things in play here:
- We want lots of tests
- We don't want to run all these tests before each commit

And two solutions:
- Make running tests faster (with faster VM startup, faster compilation, faster infrastructure, etc).
- Have a sanity test "suite" that everyone should run, and then rely on the buildbot to tell us if we're green on the rest.

 > Could we make this class of bug a test case within run-tests and
 > then make the infrastructure local to that bug? That way new tests
 > can be new files but we avoid an explosion in the number of test
 > cases.

Are you suggesting to not make it part of, eg, test-art-host? That works for me, as long as the tests are being run by the buildbot.

Patch-set: 1