Update patch set 1
Patch Set 1:

> If a component of a test is flaky then we should reproduce the
 > flakiness and isolate fix the flaky part of the test. The build
 > server is a poor substitute to engineering.

Not on large projects. Individual engineering does not scale.

 > 
 > We shouldn't have a single failing test and the build server
 > confirms this. Having the build server isolate a breaking test is
 > an extra requirement that I don't see as being entirely necessary,
 > an engineer can run tests locally and see the extent of the
 > breakage.

An engineer that was on the project for a long time, sure. But even for him, it removes the extra burden to isolate the breaking test. Isolating is the kind of automation you want from a build/test server.

Patch-set: 1