Update patch set 1
Patch Set 1:

Dmitry, Sergui: It would now be possible to "merge" the code of GenInstanceofCallingHelper, so that x86 doesn't have its own code anymore. The special Setcc code could be exposed as the new GenSelect with DCHECKs for values of only 0 and 1. GenInstanceofFinal is complicated because of potentially having to allocate a register Set-able.

Patch-set: 1