Update patch set 2
Patch Set 2:

It seems dx can generate this at the moment (at least the CTS test did this). Also, bdc wants to tolerate this as we don't know what people did before to split apps up. And yes, this can lead to failure when the classes are not identical.

Patch-set: 2