Update patch set 7
Patch Set 7:

CY,

I actually agree with Hans on most of his points. Please make an NTStoreStore so that we maintain the nop semantics of StoreStore guaranteed by x86. However, since we are using non-temporal moves, we still want to make sure we have capability to insert a store barrier.

Regarding lfence, it was basically a generalization recommendation to accommodate other architectures. But Hans is right, we have no use for it. You might want to leave it from that enum. Since the extended MIR will have field that takes a barrier type, if there is ever future need for it, it can be added trivially.

Because of this, simply add NTStoreStore for store barrier and keep AnyAny for full barrier. In implementation, NTStoreStore will generate sfence and AnyAny will generate mfence.

Patch-set: 7
Label: Verified=0