Update patch set 1
Patch Set 1:

Stack walker crashes if Proxy constructor of proxy class is on the stack. I doubt I can create a regression test for this issue due to I do not know how to make stably the case when Proxy constructor is on stack and trigger the stack walker at this moment.

Patch-set: 1