Update patch set 3
Patch Set 3: Code-Review+1

Fixed the unit test to match the new MIRGraph structures (no more GrowableArrays). Everything passes with this :) Ian, could you PTAL and consider this for merge? This is functionality we really want and it would really help to have this in common code :) Thanks!

Patch-set: 3
Label: Code-Review=+1