Update patch set 3
Patch Set 3:

> > > > Is there a performance impact with this change?
 > > >
 > > > Are you worried about moving from macro to function? If it's
 > > > properly inlined (and the compiler should warn if it can't),
 > > there
 > > > should be no difference.
 > >
 > > Not the macros, more about the extra method call between
 > Execute...
 > > and SpecializedExecute... function.
 > 
 > Ian's note to me was that as it's four parameters, on most
 > platforms that will at least not transfer through the stack, but
 > registers.

I doubt that profiling would show anything from this change. The calls are all trivial tail calls and the sole caller of a function. The compiler should be able to do the right thing, which isn't the same as saying it will - so point taken, we should watch the benchmark metrics. As Andreas points out I changed in PS3 to not pass the template parameters to the interpreter implementations and instead to just recompute the in the implementation. The bulk of the interpreter time should be going in the goto code, here there is a change to make the alt version not use a computed goto (which should be faster) and a change from macros to functions (which should be performance neutral but an aid to debugging). Sadly the template specialization change was the single biggest win in trying to fix the compile time issue. I note clang/LLVM has a lot of compilation time issue bugs open.

Patch-set: 3
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0