Update patch set 1
Patch Set 1:

(1 comment)

Thinking about how we handle this, aren't we doing it wrong, on a deeper level? I believe something like this compiles:

...<init>V
  .registers 1
  const v0, 0
  return-void

At *any* point we overwrite an (or better, the last alias of) uninitialized reference in the constructor we know we failed to call the super-constructor, and should fail.

Opinions?

Patch-set: 1
Label: Code-Review=0