Update patch set 1
Patch Set 1:

Yes, it looks like that code sequence gets past the verifier.

I'm not sure about the best way to fix it, though. How about the following?

When you're verifying an instruction (round about line 1555) in a constructor, scan the register line for an UninitializedThis register, and if you see one set a flag. End the end of the big switch statement (round about line 2771), if the flag is set, you check that there is either an UninitializedThis or an InitializedReference in the register line, and complain if you can't find one now.  The rest of the patch remains unchanged (well, apart from extracting a small function as suggested.)

Thoughts?

Patch-set: 1
Reviewer: Gerrit User 1053354 <1053354@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0