Update patch set 3
Patch Set 3:

(1 comment)

I liked initial design better where NCE was a separate pass instead of separated in type inference and instruction simplifier passes. Also, it is a bit of a stretch to have nullability analysis be part of type analysis pass - the connection I see is that reference types might need null check but not sure if this is always true. However, obviously if Nicolas is okay with this design, I can live with it :)

Patch-set: 3
Reviewer: Gerrit User 1013855 <1013855@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0