Update patch set 5
Patch Set 5:

> Looks good. :) Does the inliner mark methods as leaves if all calls
 > are inlined? Are calls to intrinsics that have implementation make
 > it so that caller is considered leaf?

The inliner is not involved here, it's just about what's in the graph and the LocationSummary of instructions.

About the intrinsics, we need to change the unconditional return true for NeedsEnvironment on the invoke call to tell whether it truly needs it.

Patch-set: 5
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0