Update patch set 1
Patch Set 1: Code-Review+2

We should probably check that the callers also are passing in Handles. Usually for stuff like this I would change the function to a static Class::GetStatus(Handle<...> ) so that callers need to pass in a handle. I suppose this is fine for now.

Patch-set: 1
Label: Code-Review=+2