Update patch set 2
Patch Set 2:

> +enh, srhines, wangying: Any good idea how to do a decent host
 > build? Would -march=native -mtune=native be possible in our build
 > system?

We could add flags to do more on the host side, but it isn't always the case that the system generating the host tools is the one running the host tools. We can't alter the tools that the SDK is intended to use because those host tools ultimately end up on our developers' machines, which may not have SSE 4.1, etc., and thus we wind up with our baseline prescott features again. What is the motivation for adding more host flags? Is it faster builds/dexopt/etc.? Perhaps this can be quantified.

Patch-set: 2
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0