Update patch set 1
Patch Set 1: Code-Review-1

The goal is to store pre-compiled odex files compressed to save space on the system partition, then decompress them into /data when they are relocated?

Out of curiosity, how much space do you save by compressing an odex file? I thought it was already a fairly compressed format.

You will need a change to the oat file assistant to be able to check the status of compressed odex files. Rather than treat compressed odex files as a separate kind of input to patchoat, if the OatFile class could be changed to open a compressed file directly, presumably you wouldn't need to modify patchoat, the oat file assistant, the frameworks, or the installer code.

In the case of PIC code, this means you would end up having two copies of the oat files, the compressed one in /system and the uncompressed one in /data?

Patch-set: 1