Update patch set 4
Patch Set 4:

Nicolas, here is as I understand the situation. On x86 we met a bug that long-to-float produces incorrect result for some values. I've upload a fix with test showing the issue. The fix has been merged but it looks it appears that arm produces the incorrect result in both quick and optimizing compiler. Merge has been reverted. This patch set restore the fix for x86 platform but temporary disables the test for arm. When the fix will be available for arm this disabling should be removed. Rolland may correct me if I'm wrong but it is how I see the situation from my side.

Patch-set: 4