Update patch set 4
Patch Set 4:

> > Nicolas, here is as I understand the situation. On x86 we met a
 > bug
 > > that long-to-float produces incorrect result for some values.
 > I've
 > > upload a fix with test showing the issue. The fix has been merged
 > > but it looks it appears that arm produces the incorrect result in
 > > both quick and optimizing compiler. Merge has been reverted. This
 > > patch set restore the fix for x86 platform but temporary disables
 > > the test for arm. When the fix will be available for arm this
 > > disabling should be removed. Rolland may correct me if I'm wrong
 > > but it is how I see the situation from my side.
 > 
 > But do we have a fix for ARM? Could we just mimic what Quick is
 > doing instead for both x86 and arm?

Actually, the current CL makes Optimizing mimic Quick.  The ARM implementation of long-to-float (and long-to-double), lacking some precision, is the same as Quick's.  We used to have a similar implementations for long-to-float and long-to-double on x86 with similar issues; Serguei's CL changes these implementations to match Quick's (by making use of x87 FILD/FISTP instructions).  The "fix" I applied was to ignore the newly added test case on ARM, which would have failed anyway without this CL.

I'm currently trying to check whether we can improve the precision of long-to-float (and long-to-double) on ARM, maybe by calling the runtime; we have an art_l2f builtin that delegates the conversion to the compiler, which itself (at least for g++) delegates to its own runtime (call to __aeabi_l2f).

Patch-set: 4