Update patch set 3
Patch Set 3:

PS3 is based on PS1. It layouts such that eh_frame_hdr is after eh_frame, that eh_frame is page-size aligned, and that the PT_LOAD covers eh_frame and eh_frame_hdr.

This is loadable for me and runs somewhat, but I get a segfault in the CFI test.

Patch-set: 3