Update patch set 5
Patch Set 5:

I also agreed with Roland to merge this CL. I would still remove incorrect value from the test but I want to see behavior corresponding to Java/IEEE specification more that cosmetic issue with test. So I think patch should be merged. Alternative way I see is to update test to be precise and just exclude the test for ARM 32-bit. IMHO it is better than we are doing now.

Patch-set: 5