Update patch set 3
Patch Set 3:

The doc you've linked to doesn't state anything about this particular case but the JLS does : "Positive zero and negative zero compare equal; thus the result of the expression 0.0==-0.0 is true and the result of 0.0>-0.0 is
false. But other operations can distinguish positive and negative zero; for example, 1.0/0.0 has the value positive infinity, while the value of 1.0/-0.0 is negative infinity."
(section 4.2.3).

Patch-set: 3