Update patch set 3
Patch Set 3:

I agree that it is odd to be different, which is why I thought the original revert was 'interesting'.  On the other hand, the code is quite explicit (and showed my bug).  I suspect it is to get ordering consistent for Doubles.  Otherwise the Double(0.0) and Double(-0.0) could end up sorted in either order using some sorts.

Patch-set: 3