Update patch set 2
Patch Set 2:

> Q: Any oat-file-assistant change?
 > A: I would suggest to not change oat_file_assistant, because it
 > needs to extract gzipped odex file to read its content. But that is
 > needed only when odex is not present and dex2oat is triggered. This
 > dex2oat command is changed to patchoat by installd if the *.odex.gz
 > file exists. In this case dex2oat could not succeed in case the dex
 > file is stripped (contains no class files). In other words the
 > gzipped odex support is needed only for precompiled odex files and
 > it would make the oat assistant even more comlicated.

Is the expectation that the following flow will be used?
1. The package manager calls getDexOptNeeded.
2. Art says: "I see no oat file and no (uncompressed) odex file, so yes, DEXOPT_NEEDED."
3. Package manager tells installer to run dex2oat.
4. The installer checks and finds a *.odex.gz file, and runs patchoat on that instead of dex2oat on the dex file.

This assumes the odex.gz file is not out of date, right? Is that a valid assumption?

This change is not consistent with the current design in which getDexOptNeeded is used to determine what needs to be done to generate up-to-date code and in which DEX2OAT_NEEDED means dex2oat should be run. If compressed odex files are going to serve as another place to look for compiled code, the oat file assistant should be updated to reflect that.

Would it work instead to support transparently loading odex files that are compressed? Then no changes would need to be made to the oat file assistant, patchoat, or installd. This may be less complicated to implement and maintain than adding another location for compiled code to the oat file assistant.

Patch-set: 2
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0