Update patch set 2
Patch Set 2:

The flow is right. But odex.gz must not be out of date because there is no way to re-dex2oat apk/jar files if they are stripped. In other words, if we have precompiled boot and application classes for a system, the apk and jar files do not have their classes.dex files. That is why I suggest a simpler way of handling odex.gz files.
I agree that transparent loading would be a better solution. The best way to implement it is to enable file system level compression like it is done in NTFS. That is not our case.
Now, reading ElfFile needs a file descriptor that is used to map the uncompressed data to memory. That is why we need a file that contains uncompressed data. If an odex file is compressed then the patchoat or libart has to create a new file that is not always possible because of some (security?) reasons. That is why the swap file is created by the root privileged installd rather than by the dex2oat itself. The improvement would be to redesign ElfFile to allow in-memory ElfFile (without file descriptors).

Patch-set: 2