Update patch set 4
Patch Set 4:

> Uploaded patch set 4.

This uses a kCfg... flag in runtime/globals.h. I think the code is clean enough that we don't need to use `dlopen()`. What do you think?

Also, I restored support for disassembly on target when not in production mode. We think it will be useful. Any arguments against that?

Ideally, I'd like to take some time a little later to try to clean-up argument passing for the printing mechanisms.

Patch-set: 4
Label: Verified=0