Update patch set 1
Patch Set 1:

I have looked thought the stdio log of the build step on server, and I worried it might be using the non-debug version.
In any case, better be safe then sorry and add the argument explicitly if we need it.

With the added printfs the error message is:
+Can not find int java.util.Arrays.binarySearch(java.lang.Object[], int, int, java.lang.Object, java.util.Comparator) in backtrace:
+  __rt_sigsuspend
+  pause
+  Java_Main_sleep
+  boolean Main.sleep(int, boolean, double)
+  boolean Main.bar(boolean)
+  int Main.compare(Main, Main)
+  int Main.compare(java.lang.Object, java.lang.Object)
+  
+Unwinding other process failed.

So if the problem is not in non-debug dex2oat, then I still do not know what the trouble might be.

Patch-set: 1