Update patch set 5
Patch Set 5: -Code-Review -Verified

We just had a discussion with Nicolas and David and we think merging this change now could be cumbersome if we need to cherry-pick later ART changes to other branches (especially in the context of the M release).

Could we defer this change?  A temporary workaround is to have all the new arch-specific instructions in the same place (nodes.h).  We can split this file later (after the release).

Patch-set: 5
Label: Code-Review=0