Update patch set 5
Patch Set 5:

> Humm, why do we need this? Can't we just consider nodes.h to be the
 > common and move everything specific out of it?

I explained in a previous comment that I took the simple big copy-paste for now so you could say how you'd prefer to have this implemented.
The dependencies are
(HGraph related code) -> (common HInstructions) -> (arch-specific HInstructions) -> (HGraphVisitor related code)

So in some way we need to 'insert' the arch-specific HInstruction definitions after the common HInstructions and the HGraphVisitor related code.

The way to do it to minimize the changes in `nodes.h` is to #include the `nodes_<arch>.h` files in the middle. Otherwise we need to do some header refactoring, for example to create another `nodes_include.h` file that includes `nodes.h` and `nodes_<arch>.h` files followed by the HGraphVisitor related code (moved out of nodes.h), and use it everywhere instead of `nodes.h`.

Patch-set: 5