Update patch set 1
Patch Set 1: Code-Review+2

Thanks for adding this.

I am not sure if you are aware/seen the checker tests :). We have a way to make this kind of testing more easier/readable/maintainable. Let's chat offline if you don't know how they work.

We prefer to have checker tests for this kind of features also because you can write more complex scenario and you see how they correspond to real Java code. We go back to gtest when there's something that's not easily testable when looking at the graph or which would require a lot of extra information to be added to the cfg.

for licm check : 445-checker-licm

That said I'm also happy to see even more tests so +2 :)

Patch-set: 1
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2