Update patch set 3
Patch Set 3:

(8 comments)

>The new interpretation of --patched-image-file is only meaningful for patching oat files, not image files, right?

Right, I made it an error to try to use it with this.

I tried to make a single patch function but the way the lock auto-analysis works keeps me from doing it in a reasonable way. I could reduce it to only 2 functions but since this patch no longer touches either of them it wouldn't make much sense to in this patch.

Patch-set: 3
Label: Verified=0