Update patch set 1
Patch Set 1:

> > You should put the explanation into the commit message to explain
 > > why this change is needed.
 > 
 > also, can we have more info on what these devices are doing
 > differently? are they lowering the stack size?

Basically, the stack decrements down to an non-allocated region (between two heaps), it triggers SIGBUS when reading/writing the stack.  The backtrace shows many DoDFSPreOrderSSARename() to DoSSAConversion() that is frame #00.  I am not sure why this happens on some of the devices.  I will try to get more info.  Thanks!

Patch-set: 1