Update patch set 7
Patch Set 7:

> > If all of the DexCache arrays are native, is there any reason to
 > > have DexCache be a Java object? Can the entire DexCache be moved
 > to
 > > native?
 > 
 > I don't want to do that. Once I use the .bss section for those
 > arrays, I don't want to immediately dirty some of the .bss page(s)
 > by writing the DexCache(s) itself. I'd rather keep these objects in
 > the Java heap.
 > 
 > That's also the reason for not using the LengthPrefixedArray<>.

I don't understand. Do you have to write the DexCache object to the .bss section? Is it because you need to have the DexCache object somewhere for the boot image, and you can't just allocate a new DexCache object on the fly that wraps the arrays in the .bss section?

Patch-set: 7