Update patch set 7
Patch Set 7:

> Yes, DexCache could be moved to native if desired. But I don't need
 > that for having PC-relative loads from the arrays and I don't see
 > any other reason to do that. There would be also the small
 > disadvantage of increasing the size of the Class by another 4 bytes
 > (HeapReference -> native pointer), so I'm not inclined to do it
 > without some other incentive.

The incentive I had in mind was that we could remove DexCache.java, and then any future changes to the DexCache can all be done in art in C++. That would make it easier to experiment with different approaches to the DexCache. Though probably not enough to outweigh the extra 4 bytes per class, you save memory from not having a DexCache class object and from not having the extra few bytes of object overhead for the DexCache instances.

Mathieu, for class unloading, is the value of having DexCache be a Java object that you can use the garbage collector to track when there are no more references from class objects to a specific DexCache?

(I don't care that much about making DexCache a native object. It just seems a little funny to have a Java object that is always accessed through JNI, and I wanted to double check it was a conscious decision instead of due to no-longer relevant legacy reasons)

Patch-set: 7
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0