Update patch set 1
Patch Set 1: Code-Review-1

There are still questions left.

The commit message isn't updated. And vmarko is the person who should review this, most likely. Plus this is the legacy compiler. Why should this be merged into AOSP, where the function should be compiled with the Optimizing compiler? (For AOSP it doesn't matter if it doesn't install on a 5.0 phone...)

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1