Update patch set 12
Patch Set 12:

> I don't see any practical difference between calling
 > DebugFrameOpCodeWriterForAssembler::ImplicitlyAdvancePC() and
 > calling CodeGenerator::CompileInternal()/CodeGenerator::RecordPcInfo().
 > All of them sample the amount of data in the assembler buffer by
 > calling CodeSize() and use it as PC and store the value somewhere.

The sampling of CodeSize() is wrong if you issue IncreaseFrameSize() (wia Push()) or DecreaseFrameSize() while overwriting_. The correct position is overwrite_location_ but the CFI writer would not cope with it as it assumes the location is monotonic. So we will need to reshuffle the emitted CFI data, not just add the "advance location" records.

Patch-set: 12