Update patch set 1
Patch Set 1:

> 1) that's not what we have here before this patch. The existing
 > response to invalid utf8 is arbitrary and unpredictable and
 > includes such possible attack vectors as overrunning the input
 > buffer and enabling many different "utf8" inputs to map to the same
 > utf16 output.
 > 2) I very much doubt this is the only or even worst thing that
 > depends on the contents of OAT files being sane.
 > 
 > If that is considered a real risk then this entire module needs
 > serious hardening, regardless of this particular patch.

Overrunning the input buffer can only cause a segmentation fault and the utf16 collisions are unlikely to cause much harm either. On the other hand, overrunning output buffer can potentially lead to arbitrary code execution. So, while I think the current code could be more robust, it's not as bad as it would be with this PS1.

Not all inputs come from dex/oat files, there's also the JNI::NewStringUTF(). And I'm not going to provide extra attack vector for someone who finds out how to trick a system process to call JNI::NewStringUTF() with an arbitrary utf8 string.

Patch-set: 1