Update patch set 1
Patch Set 1:

It's not that this is broken its just that it is long since it makes a lot of allocations. (Specifically it makes one every time an interface is initialized in order to record that the initialization happened. It also has a lot of interface classes that are generated.)

Patch-set: 1