Update patch set 2
Patch Set 2:

> Test case? What triggered it?

I basically want to assert the same-value-store (must have a previous store already) or possibly-redundant-store (must be on a singleton which is an allocation so can't be null) will have their null-check eliminated already. But the same-value-store case can still take another null-check as input as a result of GVN. So it's hard to assert here. It's probably hard to hit also, and my deleted assert is a little wrong in another sense (should be DCHECK(!ref->IsNullCheck()) instead of DCHECK(!ref->InputAt(0)->IsNullCheck()) so it's not hit in the cases I tested.

Patch-set: 2