Update patch set 4
Patch Set 4:

> ping?

Hi Mark, really sorry I haven't made progress with this yet. Definitely not giving up!

Here's what I'll do to get it moving... On Monday I'll rebase and finish the changes to boolean/instruction simplifier and get the patch ready with the fallback if/then/else implementation in all of the codegens. That is pretty straightforward and we should be able to merge it quickly.

After that, you can either implement CMOV yourself and I'll happily review your CLs, or I'll do it myself later. Same for ARM/ARM64.

Patch-set: 4
Label: Verified=0