Update patch set 1
Patch Set 1:

(1 comment)

> > Can you explain why this is safe? The reference hold by the
 > handle
 > > can change...
 > 
 > Isn't the commit message enough?

It's not. You only say that we can do it and not why. 
It is not clear from the handle documentation whether or not the StackReference may be updated or just the inner mirror pointer...

Besides, MutableHandle (who inherits from Handle) can change the reference with the Assign() method. That seems to be at least one source of contention.

I think it's nice to clarify the questions before submitting.

Patch-set: 1