Update patch set 3
Patch Set 3:

> Do we need to add a flag to say there is no compiled code?

That's the VerifyAtRuntime part from the header.

 > Does the
 > oat file not already have that information?

We do not have this flag, as it was never useful. Traditionally we haven't done derived this info (see PIC, which could check some things). I agree though that there should at least be DCHECKs to confirm that the flag agrees with the oat content.

Patch-set: 3