Update patch set 1
Patch Set 1:

Should we differ this for a follow-up patch? 

We probably first need to figure out which shorties are internal to the VM and which are not so that we only enable this pass on the ones that are not (since internal to the VM we use a hard-float ABI).

Patch-set: 1