Update patch set 1
Patch Set 1:

Yes Vladimir - that should have been "defer" :-). Auto-correcting my English did not really work there :-).

Ok - I guess for now we are happy to merge this as it is. I will follow-up with patch for the stack maps that will expose some of the is_external_to_the_vm info.

Patch-set: 1