Update patch set 1
Patch Set 1:

(2 comments)

I am not actively working on this code.  This was just a couple of days (as a side project) to understand IVA and see how it could be used to replace some existing code that doesn't do as good a job of analysis as IVA.

I consider this more of a 'cry for help' to figure out how to implement these types of changes (which are important for some benchmarks) in IVA, and do it properly.  This code isn't something that I would want to check in, in particular the TypeConversion code.  The hope is that allowing IVA to support some more cases would allow more upstreaming of optimizations that use information derived from this analysis.

Note that this is NOT a proposed change to IVA at this stage.

Patch-set: 1
Reviewer: Gerrit User 1036869 <1036869@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0