Update patch set 8
Patch Set 8:

I don't think there is a single bad case here.  That is just the one that was covered by the test suite.

Whenever we have a number that looks like

2^n-0.5-epsilon

there seems to be a strong risk that when we add 0.5, the resulting number will be rounded to 2^n by the addition, and we will get the wrong answer.

Since this probably depends on IEEE odd/even rounding conventions, and it's only an issue for exponents in the right range, my best guess is that there are 26 or 27 bad examples for double.  But don't quote me on that.

The bottom line is I don't believe this approach works.

It might make sense to try adding 0.5 and fail back to the general routine if the exponent changes as a result.  But that requires more thought.

Patch-set: 8