Update patch set 1
Patch Set 1: Code-Review+2

Would probably be worthwhile to clean up the ancient arm code that's using old syntax (some of which was written back in 2007).  For reference, here's a clang bug report describing the issue:

https://llvm.org/bugs/show_bug.cgi?id=24350

I'll file a feature request against myself to update the old arm code, and we can back out the special case when that goes in.

Patch-set: 1
Label: Code-Review=+2