Update patch set 2
Patch Set 2:

Does it matter what gcc emits?

Is it clear that xchg is worse for an atomic store, since that also subsumes the store itself?

This is starting to look really ugly.  I think the threshold for overruling the compiler code generator like this should be pretty high.  It's not that we have a different goal from the compiler here; we just disagree how to best do it.

Patch-set: 2